requirements.txt
streamlit
pdfplumber
docx2txt
groq
python-dotenv
reportlab
